Make sure sampling and event evaluation occurs after state machine update, using current frame DeltaTime
