--- Runtime/Systems/RaiseAnimationEventsJob.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 4ea173b549d44cb7aa52bc293116d40f
-timeCreated: 1655911120
