--- Runtime/Systems/SyncBlendParametersJob.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: cf27082d2bf14c4da8931a8049d6a160
-timeCreated: 1655904211
