+++ Runtime/Systems/UpdateStateMachineJob.cs
@@ -10,15 +10,21 @@ namespace DOTSAnimation
     [BurstCompile]
     internal partial struct UpdateStateMachineJob : IJobEntity
     {
+        public EntityCommandBuffer.ParallelWriter Ecb;
         public float DeltaTime;
         public void Execute(
+            Entity e,
+            [EntityInQueryIndex] int sortKey,
             ref AnimationStateMachine stateMachine,
             ref DynamicBuffer<ClipSampler> samplers,
-            in DynamicBuffer<AnimationState> states,
+            ref DynamicBuffer<AnimationState> states,
             in DynamicBuffer<AnimationTransitionGroup> transitionsGroups,
+            in DynamicBuffer<BlendParameter> blendParameters,
             in DynamicBuffer<BoolTransition> boolTransitions,
             in DynamicBuffer<BoolParameter> boolParameters,
-            in DynamicBuffer<ExitTimeTransition> exitTimeTransitions
+            in DynamicBuffer<ExitTimeTransition> exitTimeTransitions,
+            in AnimatorEntity animatorOwner,
+            in DynamicBuffer<AnimationEvent> animationEvents
             )
         {
             AnimationStateMachineUtils.RaiseExceptionIfNotValid(stateMachine, states);
@@ -83,6 +89,32 @@ namespace DOTSAnimation
                     var nextState = states.ElementAtSafe(stateMachine.NextState.StateIndex);
                     nextState.Update(DeltaTime, ref samplers);
                 }
+            }
+            
+            //Sync parameters
+            for (var i = 0; i < blendParameters.Length; i++)
+            {
+                var blend = blendParameters[i];
+                var stateIndex = blend.StateIndex == stateMachine.CurrentState.StateIndex
+                    ? stateMachine.CurrentState.StateIndex
+                    : blend.StateIndex == stateMachine.NextState.StateIndex
+                        ? stateMachine.NextState.StateIndex
+                        : - 1;
+                if (stateIndex >= 0)
+                {
+                    var state = states[stateIndex];
+                    state.Blend = blend.Value;
+                    states[stateIndex] = state;
+                }
+            }
+            
+            //Raise events
+            RaiseStateEvents(sortKey, stateMachine.CurrentState, animatorOwner.Owner, states, samplers,
+                animationEvents);
+            if (stateMachine.NextState.IsValid)
+            {
+                RaiseStateEvents(sortKey, stateMachine.NextState, animatorOwner.Owner, states, samplers,
+                    animationEvents);
             }
         }
         
@@ -155,6 +187,36 @@ namespace DOTSAnimation
 
             nextStateIndex = -1;
             return false;
+        }
+        
+        [BurstCompile]
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        private void RaiseStateEvents(
+            int sortKey, in AnimationStateMachine.StateRef stateRef,
+            Entity ownerEntity, in DynamicBuffer<AnimationState> states, in DynamicBuffer<ClipSampler> samplers,
+            in DynamicBuffer<AnimationEvent> events)
+        {
+            Assert.IsTrue(states.IsValidIndex(stateRef.StateIndex));
+            var state = states.ElementAtSafe(stateRef.StateIndex);
+            var samplerIndex = state.GetActiveSamplerIndex(samplers);
+            
+            var sampler = samplers[samplerIndex];
+            var normalizedSamplerTime = sampler.Clip.LoopToClipTime(sampler.Time);
+            var previousSamplerTime = sampler.Clip.LoopToClipTime(sampler.Time - DeltaTime * sampler.Speed);
+            
+            for (var j = 0; j < events.Length; j++)
+            {
+                var e = events[j];
+                unsafe
+                {
+                    if (e.SamplerIndex == samplerIndex && e.NormalizedTime >= previousSamplerTime &&
+                        e.NormalizedTime <= normalizedSamplerTime)
+                    {
+                        var ecb = Ecb;
+                        e.Delegate.Invoke(&ownerEntity, sortKey, &ecb);
+                    }
+                }
+            }
         }
     }
 }
